backend/requirements_backend.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
